xb1: Change default HDCP protection mode
This CL updates default HDCP protection mode from HdcpProtection::On to
HdcpProtection::OnWithTypeEnforcement.

It seems like HDCP protection setting mode On always seems to work.
This includes when the output is a VGA cable.  However, this behavior is
not desired.

On the other hand, OnWithTypeEnforcement has been known to produce the
desired behavior with certain TVs.

In addition, FilterBasedPlayerWorkerHandler is modified to take this
change into effect.

b/65117856

Change-Id: I64490098b42992aa8887ae8d487293b9d66a338b

Former-commit-id: c8cee0b8b96dd92436dc70bcef5994236d4e8c58