New unit test runner implementation.
This is a first pass at the new Starboardized unit test runner, which
makes use of the new app launcher system.  When a new test binary is
added, it should be added into the master list in test_targets.py. Platforms
can filter out tests via a GetTestFilters method implemented in their
PlatformConfig class located in a gyp_configuration.py file.  Further
information can be found in the README.

Change-Id: I591d20aea984b50664e8f37234c4f77600162cb5

Former-commit-id: 65b134b035f9f283dfcae75b12a6214e2267abd7