Remove NOTREACHED() from WebDriverServer
Previously, the WebDriverServer would crash Debug and Devel if it
received an HttpMethod that wasn't recognized. This crash was getting
hit from messages being sent from the wild (perhaps security scanners).

The issue is a bug between InvalidCommandMethod() and
HttpMethodToString(). The InvalidCommandMethod() knows that the
requested method is invalid, and calls HttpMethodToString() to retrieve
the string for an invalid method; however HttpMethodToString() treats an
invalid method as an unreachable case and DCHECKS on it.

The DCHECK has been removed, since this is a valid case.

This fixes b/67047722.

Change-Id: Ia7f769650efc02da58e75ec5530391de5285dfea

Former-commit-id: 0e3c8cac13a748204165a3d03c9ffaacd8aaca58