Fix webdriver script execution failure
This change fixes the webdriver script execution failure due to
its parameters object getting garbage collected before the function
was actually executed.  The fix is implemented by preventing garbage
collection right as the Wrappable ScriptExecutorParams is created,
and having its client call AllowGarbageCollection when it is done.
This is required in order for the ScriptValue::Reference that
references the |function_object_| that it holds to actually get
traced.

b/66996907

Change-Id: I3f9052c0537916da5eea08ff45a2a53a6d9ca7b5

Former-commit-id: a49b2f419d8d2bb84e8017a03ad8b615496fc52a