Updates gyp_configuration.py files for new test runner.
In order for the new test runner to work with all platforms,
their PlatformConfig classes need a method called "GetTestFilters".
This change adds that method to all necessary files.  In most cases,
modifying a shared gyp_configuration.py file was enough, but some
platforms had special cases.  There are stub methods in platforms
that currently do not have unit testing to allow support for it later.

Roku has been excluded because they own their Starboard port, so for the
time being we are not going to touch it.

This change also splits all of the test filters from
"src/tools/lbshell/project_configuration.py" into the "GetTestFilters"
methods, effectively Starboard-izing unit tests.  The only thing that
was not carried over from "project_configuration.py" was a variable
called _DISABLED_ASAN_LEAK_DETECT_ENV_SETTINGS.  This variable involved
the use of environment variables, which the new test runner and launchers
do not support. After speaking with mkf, the decision was made to leave this
out for now.

Change-Id: Ie921a640e94b104374bbaae7dcb34d9d8f739fc2

Former-commit-id: de078c3f4ff29b4a50ae3e878eefd335ea95b3c1