Add Cobalt V8 shell
This change implements the necessary subset of cobalt::script to get
a Cobalt based shell on top of cobalt::script and V8 up and running.

Also update a few parts of documentation regarding JIT, to
accommodate the fact that V8 can only be run with JIT.

Also stub out a few bindings components in order to be able to
successfully build.

b/67313140

Change-Id: Ibc1bc33319b24edfefa101cb09619b0e50900e13

Former-commit-id: 8e6a678f5b478750e0876da03e953a90fb9fcdaa