Updates gyp_configuration.py files for unit tests.
This change adds a "GetTestFilters" method to the PlatformConfig
classes within Starboard platforms' "gyp_configuration.py" files.
Often times, modifying a single shared PlatformConfig class was
enough, but there were a few special cases.

There are stub methods implemented for platforms that do not run
unit tests.  I have omitted Roku, because they own their Starboard
port, meaning that we shouldn't touch it for now.

The first time I tried to do this, a typo that was not caught by
try builders caused the x64directfb-future builds to break.  This
second attempt should fix that; I was able to run a
x64directfb-future build on my machine with no problems.

Change-Id: I61a7a4bde78fea90ce4c0cb337cb02e8e56d5b0b

Former-commit-id: 83f27a70198357b77d6e32632be88bc4016d32db