Add ability to disable the splash screen on reloads
The raspi-2 when using the Skia rasterizer is unable to play videos
after loading the splash screen several times. Presumably, this is
because it creates an extremely large offscreen surface each step
in the animation and raspi-2-skia has a poor interaction with this.

This is breaking the webdriver benchmarks on raspi-2-skia, because the
benchmarks reload Cobalt many times; a command line parameter that
disables the splash screen on reloads is being added: 'disable_splash_screen_on_reloads'. This will allow raspi-2-skia to
record startup numbers on launch, but still be able to play videos
later.

Change-Id: Ia524b2452a7ab96793b4f9c3e493ed1455da3072

Former-commit-id: 32d7a778bb613fc55e3f45d523ff3945202a1f97