xb1: Add ability to run unit tests on the Xbox
This CL adds the ability to use Microsoft's Device Portal API, to enable
automated installation and running of binaries.  The output of the unit
tests can easily be viewed by the developer on their command prompt.

It tries to integrate with the abstract launcher python interface,
so that less maintainence is needed as we move forward.

b/38114877

Change-Id: I51d3f0f3fae1e8d6bbd1d5632db02940a7c5fbfd

Former-commit-id: 1a61ee5d7424549b379c31c1e92006f38e281747