Use top and bottom to calculate containing block height
Previously, the logic in Cobalt assumed that the height of a containing
block was not explicitly specified if it did not have a non-auto height
property. However, the height can also be specified if the top and
bottom properties are set. Cobalt now treats the top and bottom
properties being set as an explicit height of the containing block,
although height is given priority when it is set.

This mimics the behavior of Chrome and Firefox and fixes b/33924319.

Change-Id: I72ecc16a27349d7ef0914949fa0ce77cb4251d9b

Former-commit-id: 271ebdd6ed08ea9ef93dd36d38ef7afc3493e60a