Make map-to-mesh demo web page work on the PS4.
The demo web page contains only a single video element.  On the
PS4, we have a fallback splash screen which will be used here.
After rendering it though, the GL state will be left with
glBindFramebuffer() bound to a non-default framebuffer, and so
since all subsequent renders to not invoke skia, it never gets
reset and the 360 video renders do not appear.

This explicitly resets the bound framebuffer each frame.

b/67534517

Change-Id: I8802b87d2a7412de97e2494a0e902191e0dc6552

Former-commit-id: e9cb3faa9af226181596e33a7b6db739d60602dc