Make AudioRendererImpl use AudioRendererSink
AudioRendererSink is a sink class with similar interface as SbAudioSink.
Using it inside AudioRendererImpl instead of using the SbAudioSink C
interface directly allows us to replace the audio sink used by
AudioRendererImpl with implementations not based on SbAudioSink.  This
makes testing easier.
Also implemented AudioRendererSinkImpl on top of SbAudioSink.

b/67587022

Change-Id: Ie3336f3522ca674161f60bf8589e43c9202f7e04

Former-commit-id: 35508a5b940dc22fbbe619f96acce0594a2e6fe9