Load cookies without domains specified from storage correctly.
When a cookie does not explicitly specify a domain, as is the case
for the QualE site, its domain gets internally stored without a
"." prepended to it (in contrast to when the domain is explitly
specified in the cookie string).  When we save the cookie to
local storage and load it the next time Cobalt is started,
we used to pass it through CanonicalCookie::Create() again, but
this time the domain *will* be specified and so it will get loaded
with a "." prepended to the domain, and so it will no longer match
cookies coming from over the network (that won't have the ".").

This change makes it so that we just deserialize stored cookies
without applying any processing to them.

b/67624195

Change-Id: I61adb22b8dd53c0b6973413901441bc2a3848d99

Former-commit-id: f302611df2396fec34ab80ec37dbb3fdd53a369e