Fixes test runner errors
When I ran the new test runner on buildbots, I noticed several
errors with imports, etc. that needed to be fixed.  I also noticed
that Android unit tests are enabled on buildbots but cannot actually
run, so I filtered all of them out.  I have stored the known Android
test failures in a list so that we do not lose track of them before it
is time to enable automated testing for Android.

Also, I fixed an issue that caused the new test runner to try to build even
if there were no test targets specified, and added an option to disable
all testing for a platform via a single test filter constant.

Change-Id: I018b0a932d8bf5f951239ed1c1329611ba64a043

Former-commit-id: 0729123ce2d238d8b1eff30c65ac59db1b0f8df0