No longer automatically navigate on resume after error
Previously, Cobalt would always attempt to navigate after starting from
preload or resuming from suspend if an error had occurred during the
previous navigate. However, this caused issues because this navigate had
the potential to stomp on the platform response to the network failure;
this occurs on Android which has the following flow:
OnError() -> RaiseSystemPlatformError() -> Suspend() -> Resume() ->
OnNetworkFailureSystemPlatformResponse()

Cobalt will no longer automatically navigate in this case. It now fully
leaves it up to the platform as to how the error should be handled.

The only case now where a navigate happens on Resume() is when it was
initially blocked by the app being suspended when Navigate() was called.

Also, Timer::Stop() is a no-op if not running, so remove redundant checks.

b/67780587

Change-Id: I725ac5ebd7228f36a87f0e4b5b3b50a681c1e677

Former-commit-id: 1abef24247cb5fedec1d9fde1a4e0ebc91dd3191