Fixes issues in test runner
I discovered that the new unit test runner can hang if an error occurs
in the launcher that runs the tests.  This is because the reader thread
is never killed, causing it to hang forever.  This change adds exception
handling that will kill the reader thread and ensure that the test
runner's results report an error.

In addition, I discovered that you could not pass command line
parameters to the test executables.  I have added logic in main()
that takes target parameters passed in from the command line and
passes them to the launcher.

Change-Id: I88b6b0dc1a7659c8ce21611f506764d16e1a1329

Former-commit-id: d2585b8b61856ebeda6b72e691db01365c602ca1