End playback when Pipeline is ended
Previously the HTMLMediaElement ends playback when the current time is
beyond the duration of the media.  Due to round up issues the current
time may be slightly off.  The Pipeline signals the end of playback to
WebMediaPlayerImpl using OnPipelineEnded(), now forward this to
HTMLMediaElement so it can end playback in such situation.

b/64211755
b/62103445

Change-Id: I26b956247b3e075cd4171380955b3c31a0b0bca4

Former-commit-id: 606af8f45b8aee3e145e7fc961a428b3b77df782