Refactors app launcher abstraction to accept kwargs
When discussing the concept of how the launcher framework
should accept a secondary out directory, the conclusion was made
that it should be possible but not forced upon clients or implementers of
new launchers. To do this, and to handle similar cases, AbstractLauncher
and its subclasses have been updated to take kwargs.  This way, the out
directory can be passed as a kwarg that AbstractLauncher uses, but it
does so behind the scenes so that concrete implementations can assume
the path to their target will be correct whether an out directory was
provided or not.  I have updated target command line
parameters and the launcher output log file to be kwargs as well.

All launchers now get the path to their executable via
an inherited method called GetTargetPath().  It treats all platforms
as if their executable was located in the top level of their out
directory, but it is easy to use this path to get other components
in the out directory that a launcher may need. (e.g. Android).

I also quickly fixed an issue in the android launcher where some
crashes weren't being detected properly.

Change-Id: I4d55ff68211b2befae447762102a3b11d3ee12be

Former-commit-id: 9a809031eb7188f8b06f8b8a0784acf7dd4ba9d4