Handle values of type Symbol in conversion_helpers
Previously, our code would DCHECK that JavaScript value conversions
to numeric types were successful, however the ES6 spec adds the new
case of the converted from value being of type Symbol.  Since it is
perfectly valid for client JavaScript to attempt to convert Symbols
to numeric types (in which case we throw a TypeError at them), we
must add logic to handle this case.

b/37628715

Change-Id: I060e9883bed49da51775f1b8fce6b663aed4c577

Former-commit-id: 2b3ad6219b10d255e547b545f1abeae652431f4f