Allow handling of host-generated SbInputEvents.
Host-generated events could cause an infinite regress when combined
with a host-defined event handler.  Becuase YTVR's event handler
ignores SbInputEvents for right now, we punt on this problem until a
future CL.

The host is required to provide an SbWindow to associate generated
events with, so we also expose the primary SbWindow in the CbLib
interface.

b/36519720

Change-Id: I1fb254303436be7f352bb593579618af82708711

Former-commit-id: ae908a3dcdd825abd52ffe9ef794876c61783153