Change atob return type to ByteString
Previously, our atob implementation would return a
DOMString/std::string, which when passed to bindings conversion
helpers would be unconditionally converted to a JavaScript UTF16
String as if it were UTF8, which our Atob is not guaranteed to
produce.  This change changes atob's input type to the whatwg input
type ByteString, which we internally back with a
std::vector<uint8_t>, and convert as bytes to UTF16 for JavaScript.

b/67906178

Change-Id: I3ddae1bb3a51f44e78361bd83a585b11b1fe3c6e

Former-commit-id: 9564c1cccaf3cb7fe4d26a953cdff6e6af0fb526