Refactor Android audio and video decoders
There are a few hundreds lines of similar code between the audio and
video decoders on Android.  I am going to bring them as close as
possible in order to merge them eventually.  This will be done in
multiple steps in order to minimize the effort of code reviewing.

b/36923502

Change-Id: I1ddce4ae91343e847439e5b5adb8e3d08abaff37

Former-commit-id: 09ad1ed0a8b362cc58439189bbcae421422720f4