Updates launchers and test runner for env variables.
There was a feature of the old test runner/launchers that allowed environment
variables to be specified when executing test binaries.  We initially
left that out with the new system, thinking that we didn't need it,
but it turns out we do.  This change adds an "env_variables" kwarg to the
abstract launcher that concrete implementations can use to set up their
own environments.  The test runner now accesses predefined environment
variables via platforms' gyp_configuration.py files, and passes them
to its launchers.

Change-Id: I7f4bfe468db390af43ed083b9b0907e10849929e

Former-commit-id: e9967cf8a3aa4bfdd1e9eed07566b6e1078e5bb3