Fix outline animations.
Outline animations were written with the (correct at the time)
assumption that the render tree node builder object passed into
the animation callback would be the initial one.  Later, in
c236c00814, things were changed so that the builder object
modified from the previous animation callback is the one passed
into the next animation callback, and the outline animation
was not corrected when this later changed was introduced.

Update layout_tests to sanity check that re-applying an
animation with the same time does not produce different results,
to catch problems like this in the future.

Add the pointer events demo web page that ***REMOVED***@ created to
the demo web page directory.

b/68340926

Change-Id: I9327fdcaee8bc51c736f12afcfd837e1ccbd9ef8

Former-commit-id: 4aa258d1f434da63d071f845910981341405faaf