Updates requirements for common command line parameters
This change makes it so that a platform and config
are required by all tools that use the shared
starboard/tools argument parser.  In addition, it takes
the target name parameter out of that parser, because not all
tools will require a target name.  Instead, the target name paramter
has been added to the tools that need it.  Also, the CreateParser()
function will ensure that a help message is displayed by the tools
that call it in the event that no command line arguments are passed in.

b/68051454

Change-Id: I6711cf034dce2a94dc76ce5567e545ad01399e56

Former-commit-id: 83db12f9f39399c0cd6ce9fb8c61cb6f77119f89