Move CreateGlobalObject(...) implementation to member function
Previously, the templated global interface version of
|CreateGlobalObject| was implemented by having the generic version
downcast itself to the engine specific implementation inside of
abstract script.  This change refactors the general script version
to forward to an engine specific member function instead.

This has the benefit of slightly better readability in the
implementation, as well as not having to expose everything in the
engine specific global environment that would be needed to implement
the function (or the use of friends).

Change-Id: Ida93b2b9236e9ecbb423fe47e53859ae08935cbd

Former-commit-id: babd550427b52c7789bb58d12e1cb987bee10a47