Keep event_target_ alive for event processing.
It is possible that while handling one of many events that
are queued up, garbage collection runs and destroys
event_target_.  This ensures that it stays alive for the
duration of event handling.

The bug 68683453 was reproduced by manually forcing garbage
collection after each event was processed, and when doing that
a crash was observed most of the time after playing 1 video (in
particular, it was on the MediaSource event target).  This fix
was observed to solve this issue.

b/68683453

Change-Id: Ib781695c788af857b80d9c48993459b740c688d3

Former-commit-id: 0e299c25c6f5d8388f0d0845094d45bc2c49a851