Remove |OpaqueHandle| from cobalt::script
|OpaqueHandle| and |ValueHandle| are currently almost entirely
identical in functionality.  This is both confusing for clients of the
interface, and leads to duplicate logic in the implementation.  In many
cases, if the value if truly opaque, then a client shouldn't need to
care that it is an object.  IDL object conversion, which was the only
area in which these two differed, is now instead handled by a
conversion flag.

If we do need to care about whether the values are objects in the
future, then we can look into exposing that on the existing values,
or adding some new type that relates to values in a way that does not
require as much duplication.  In any case, I don't think it'd be the
current approach that we have here.

Change-Id: I164c39efdc225c8c28586f4ea8b2d20026ccdfd3

Former-commit-id: 721528bb53228e1b1fd299309ea801fe5ade6c4d