Fixes DCHECK failing where OnThreadShutdown() wasn't called.
All threads need to be detached before they exit. Originally, all native
threads were detached in the pthread local key value destructor
Destroy(). This behavior was then updated so that native threads are
detached in JniEnvExt::OnThreadShutdown(). For native threads created with
SbThreadCreate(), OnThreadShutdown() is called from thread_create.cc before
signaling to pthreads that the thread is completed. Destroy() was updated to
a simple DCHECK that verifies that the thread was already detached.

Before:
The background/main starboard thread isn’t created with SbThreadCreate()
like the other native threads are, so it was never calling OnThreadShutdown()
before it exits. Thus causing the DCHECK in Destroy() to fail for this thread.

Now:
The background/main starboard thread calls OnThreadShutdown() from
android_main() before it exits and the DCHECK in Destroy() now passes.

b/65205427

Change-Id: I915503c831d838fcf45dc136117fbdadf2484870

Former-commit-id: 17b9a291f89f5940989504be33ebc84cc3740e48