Always render the last frame during seeking
Starboard filter based pipeline uses the audio time as the playback time
of the video and sync the video frames to the audio time using the
timestamp of the video frames.  During seeking, the audio renderer and
video renderer will be independently seeked to the required time.  If
AudioRenderer::Seek() is called before VideoRenderer::Seek() is called,
it is possible that the frame returned by the VideoRenderer is not the
last displayed frame at the previous playback time as the retrieving of
the current video frame can happen on another thread.

Seek video renderer first can ensure that the last displayed frame is
always displayed.

b/68793315

Change-Id: I36a597046d233e8e22761d9d4a33270450277734

Former-commit-id: 60177f8bfc3a80de5bc00c1e4531bc4a77c534cf