Add command line parameter for retaining web fonts on suspend
To give platforms the option of retaining web fonts on suspend, rather
than always purging them, a new command line paramter has been added:
retain_remote_typeface_cache_during_suspend. When this is included, the
typeface cache is retained on suspend.

Additionally, ResourceProvider's now have a GetTypeId() function. This
allows the WebModule to compare the old type with the new one when the
resource provider is set and to purge its caches if the type is found to
have changed.

This provides a workaround for b/67672824.

Change-Id: If97d27cd5dc474433c121189d7eb5406677044a4

Former-commit-id: c0e14a987a8b82fa7e53e5176bab0ee36e2c48c1