Fix V8 deprecated API usage issues
V8 supplies V8_DEPRECATED and V8_DEPRECATE_SOON macros, that use
GCC/Clang __attribute__((deprecated)) to emit a warning when one
attempts to call a function that is marked deprecated.  Unfortunately,
we are unable to enable this, as we build with "-Werror", and a fair
amount of our code is already a client of deprecated APIs (such as
net/'s usage of ICU, for example).

For now, locally enable this setting, and fix all issues found.  In the
long run, we should look into seeing if we can disable this warning in
as few areas as possible, as opposed to globally.

b/67313140

Change-Id: Ia19196dc3c479941364dc3df68504b077a26a539

Former-commit-id: 8c47852040c37544f240c1054cbda2264977a98f