Add Preload webdriver benchmark test
There is now a Preload webdriver benchmark test that inherits from
Sanity. It validates that it is in a preload state, sends a resume
signal to Cobalt, validates that it has left the preload state, and
then runs the Sanity logic.

Additionally, the test runner has been set up to support sending
resume to the launcher.

Lastly, test.py is set up to add a preload commandline parameter
in the case that it is a PRELOAD_TEST.

Change-Id: I18e170edda66dcae26d400b7f0dd9ea4764bbcb4

Former-commit-id: ddf57b3db6d66a1c4966db2cd661d582d558955a