Fixes hang and output cutoff in new test runner
If the launcher running the unit tests errored out before it
could close the pipe it was writing to, the thread
in the test runner reading from that pipe would hang.  That
should be fixed now.

In addition, when trying to run unit tests on Android, some
output would be cut off from the buildbot logs, even when the
TestLineReader had read them all.  I fixed this by flushing
stdout.  Not entirely sure why this only happened on one
machine, but if it happens on one it could happen on more.

Change-Id: I1a0bae7791a1c543b13b59e3fc406ac932f431a2

Former-commit-id: ef12f52711df0f3b30c8b1f8ec9b0ba6f83ec8d0