Change misleading names in WeakHeapObject
GCThing is already a concept in SpiderMonkey.  Instead of being unhappy
with their definition, use the same as theirs, and just be explicit
that we don't want to track null or undefined values.  Also, null and
undefined are distinct JavaScript values, so don't use null (previously
in IsNull) to refer to both of them.

Change-Id: I0599c9cc1492c1908d2d8ed2d47f6b0511de53e6

Former-commit-id: e0f3496a95f24d93cb1bf4f9037b2591a2917d40