Make sync_to_build_id clone missing repos.
With this change we can sync back to a moment in time when
a repo that is now removed was still available.  In this
case, it will clone that repository first before setting
it to the appropriate hash.

For example, this helps with:
sync_to_build_id.py 68337

where the directory third_party/mozjs is now removed.

Change-Id: I3e91438f3c55cabcc1fbaba9b47f1288f8bd9ec4

Former-commit-id: 611a16a929e8202b6ccd1df8a61059e9d08a4ed3