Unify AudioRenderers and VideoRenderers
The media stack used to have a single implementation of AudioRenderer
and VideoRenderer.  The two classes are later abstracted into interfaces
to allow Android to have its own logic on various media stack tasks.
Now these classes are unified and turned into concrete interface again
to allow a single implementation to be used across all platforms for
maintenance and testing purposes.

There are also the following side effects:
1. HostedVideoDecoder is turned into VideoDecoder as all frame
management is done by the VideoRenderer now.
2. VideoFrame is turned into an interface where individual platform can
have its own VideoFrame implementation to wrap platform specific data.
It is also moved into filter/.
3. Closure is replaced with std::function<>.
4. Host classes are replaced with std::function<> based callbacks.
5. The Android initial seek timeout logic on video is modified as now
the timeout is calculated from the time that the first input buffer is
written instead of from when the time that the first frame is decoded.

b/36923502
b/69713848

Change-Id: I6e8a5a293db6f8e2dd18b8bf174defedf70c1172

Former-commit-id: f926bd3a65551ca7692ab3b9ff7dd7c36989bd34