Fix logic calculating coordinates when exiting transform
Previously, there was a bug in Box::ApplyTransformActionToCoordinates(),
where it was incorrectly including the full offset from the root in its
transform rect offset when exiting a transform. Applying the full offset
is only correct when entering a transform.

When exiting a transform, only the containing block offset impacts
contained coordinates. The offset from the containing block to the
root is added to the coordinate calculation after the transform is
exited.

This fixes b/68715873.

Change-Id: I97b218c37e01b7f13f6e10f1279843bc284068a3

Former-commit-id: f1bc7e9279138905addc486346349a2e8e4a3554