Handle JS::To{Uint,Int}{32,64} failure in conversion_helpers
It is possible for JS::To{Uint,Int}{32,64} to fail in certain edge
cases, such as application JavaScript setting up infinite recursion
that gets triggered by the conversion, as observed in b/69181010.  This
change adds logic to handle this case, by removing the DCHECK on the
conversion being successful, and ensuring that an exception is thrown.

This change also adds |SetExceptionAlreadySet| to |MozjsExceptionState|,
which is important in order to get the appropriate exception to client
JavaScript.  This is required, as we must be able to handle arbitrary
exceptions being thrown from the JavaScript that was executed as a
result of attempting the conversion.

b/69181010

Change-Id: Ia64d3f67b39c9cb34aec09aa6203582827b3f3ce

Former-commit-id: c45a5e33d8c43d63fa9c2292c08772882b40d55d