Fix fallback font selection to use system language-script
Previously, there was a bug where the system language string was being
passed to fonts for font selection. However, base::GetSystemLanguage()
actually returns a language-country string, while the fallback fonts
expect a language-script string.

To resolve this, a new base::GetSystemLanguageScript() function has been
added. This value is passed to the FontCache for use when requesting
fallback fonts.

This fixes b/68771869.

Change-Id: I460a24832774ae3e1b0255c7d78faf7c9ba3a917

Former-commit-id: cf328a48e7f939fd09eabcf4f9c4f8a4c12db2c5