Convert visible_rect parameters
Previously, Rect would allow passing in int64s.  It would silently
truncate by doing a static_cast.  Going forward, a change in Cobalt is made
so that if int64s are passed into the function, they will possibly match
with the constructor with float arguments (which is deleted).  This is
on purpose to catch these cases, so that we must think through the
behavior when int64s are too large to fit into integers.

The solution chosen is here is practical, such that it will convert the
arguments passed into RectF's constructor to a float, and then a Rect is
generated from that RectF after inputs are clamped to the range of ints.

This means that if really big int64s are passed in, they will be
clamped, and if reasonable parameters are passed in (up to a value of few
million), there will be no loss in the int->float->int conversion of
parameters.

Change-Id: I19796938476450e0911c4ba06d421071798955e8

Former-commit-id: 64b8ec59d291bc3fd76dc79c5c9c4e01b4db3d78