Fix index calculation for style sheets
There was a bug in the calculation of indices for style sheets where all
style sheets in a list were being given the same index. This caused
incorrect styles to be selected during rule matching.

A test has been added to verify that this is working properly.

This fixes b/67648453.

Change-Id: Ie65e5450b424dbc374e550e6a75039096ba292dd

Former-commit-id: a8a640028b31b844970604b2f5a67fc17d755a06