Update launcher for ADB quirks.
It seems that not all versions of ADB support the syntax the
launcher previously used to list installed packages.  I verified
that this new syntax works on Crow and the Nexus Player hooked up
to the test rig.  I've also discovered that ADB commands do not
seem to be good at providing correct exit codes, so I explicity
check whether the line indicating that our package is installed
is present.

I have also moved all of the logic for checking the presence/validity of
the app's exitcode file from the main run loop to the exit file
watcher process. On the buildbot, errors from "adb shell" go to stdout
instead of stderr for some reason, and the exit file watcher can now
account for that.

I also added a check that makes sure the launcher
won't exit until all of the executable's logcat output has been read.
This is to prevent the case where the app exits and the exitcode is
retrieved, triggering the shutdown of the launcher before it can finish
reading logcat output.

And I added timers for install/run steps.

Change-Id: I34c575a3a9a8baecfeaed197fb09dfda5b0f70cb

Former-commit-id: a6f9b80a83905abfe5df3abde8cae211ab401f39