Add support for loading retries in the resource cache
The resource cache now provides the ability to retry on loading fails
when the failure comes from a transient error and the resource cache is
set to allow retries. Currently, only the RemoteTypefaceCache enables
retries.

The following NetFetcher errors are treated as transient:
5XX response codes, ERR_TEMPORARILY_THROTTLED, ERR_NETWORK_CHANGED, and
ERR_NAME_RESOLUTION_FAILED. The first three are error types that
UrlFetcherImpl is willing to attempt retries on.
ERR_NAME_RESOLUTION_FAILED indicates a socket error.

The retry scheduling itself is done as an exponential backoff. The delay
doubles on each attempt until the max delay of 1024 seconds is reached.
At that point, it will continue attempting every 1024 seconds.

This fully resolves b/67672824.

Change-Id: I08c2f1797305c7b179076277edbdad1fd8f35b5b

Former-commit-id: 5b4ee8ea9b2fb48fe33485f2949c553fe576d573