Allow H5vccAccountManager to shutdown while an auth request is pending.
Use a weak pointer to H5vccAccountManager in the message loops so that
it won't be called in case it gets destructed before the message is
processed.

Keep ownership of the ScriptValue reference to the callback in
H5vccAccountManager rather than precariously passing it to the platform
user authorizer and back.

Notify the user authorizer when H5vccAccountManager is shutting down so
that it can unblock any blocking request to avoid a deadlock when
waiting for the account manager request thread to exit.

b/68761280

Change-Id: I25fc419ba1451e91752792235462f385a2808381

Former-commit-id: bfe2513ad8fa33c2fb8b6080c7fd40c237563968