Fix crash inside ShellDemuxer::ParseConfigDone()
ShellDemuxer executes most of the tasks on the contained
|blocking_thread_|.  It destroys the thread in the dtor to guarantee
that all tasks running on the thread holding a reference to the
ShellDemuxer instance are destroyed and will never refer to an invalid
ShellDemuxer instance.

However, ParseConfigDone() is not running on the blocking thread.  It is
possible that when ParseConfigDone() runs, the ShellDemuxer instance has
just been destroyed.  Now ParseConfigDone() is posted with a WeakPtr to
ensure that it won't run if the ShellDemuxer instance is destroyed after
this task is posted but before it gets a chance to run.

Also add for |stopped_| inside ParseConfigDone().

b/69480214

Change-Id: I59a635054ee5783f0f361e2d14485213105a25a7

Former-commit-id: 874c8c7d71d27bc68a47b03f6ff86409f3c3bba5