Add "-Winconsistent-missing-override" gmock workaround
Gmock doesn't offer any way to supply an override specifier after a
method defined by MOCK_METHOD*.  This issue is tracked internally at
b/18208154.  The two options we have are to globally disable the
warning, or do what Chromium does, and comment out the other override
in the class (which will cause the warning to not be triggered).
Consider this only happens in one single place throughout all of our
code, let's go with working around gmock.

Change-Id: Ib63f79de15bef12ef2d3e7c2f21ff294eb165f92

Former-commit-id: 6a08118456a7e3d721652e548c189ce73999d235