Remove incorrect override on function implementation
Right, so this is another example of why it was bad to have OVERRIDE
defined to be nothing on Clang.  linux-x64x11-internal was the only
platform to compile this code, which used clang, where OVERRIDE was
substituted for nothing.  However, putting a real override here results
in an error, as it is not legal C++.

.../my_new_interface.cc:26:35:
error: expected function body after function declarator
MyNewInterface::~MyNewInterface() OVERRIDE {}

Change-Id: I32e8011a31e97e0a3bb87ff4abf60ea9469d0330

Former-commit-id: c0ecd5769d6d964242289b7c05bdc9e1fedc5948