win32-lib: Mirror host texture.
Allow mirroring a host-specified texture to the system window.

This is a temporary hack/fix, to work around the fact that mirroring at
the EGL/GL level is broken -- likely due to things in the host code.
The intention is to fix the host code ASAP and remove this.

The cleanup task is tracked in b/69924293.

b/67379826

Change-Id: I839ff9322978c0a55588409bd3699f4100606165

Former-commit-id: 58b129ba9ed39f7db0ecf53cc171faeb6bda37ad