Reduce code duplication and complexity around Promise
This change adds a new type, |PromiseResultUndefined|, to represent
resolving a Promise with undefined.  This allows the void
specialization of |Promise| (as well as its implementation in each
engine, |NativePromise|) to be a simple wrapper over the undefined
version, which enables us to not have to copy paste code between the
non-void and void versions.

In the future, once we have our low level JavaScript abstraction layer,
that exposes arbitrary values, we should use that instead of something
specific to promises.

Change-Id: I5a523e868c9d9da001974ad2fae65a754bc30cb4

Former-commit-id: 87778d84d12979e95f461e8be656d02037a08955