Escape command line metacharacters in flags for raspi
Previously, the raspi launcher could fail to successfully run Cobalt
when the url included & in it. That is because the launcher was not
escaping the command line metacharacters in the url before launching
Cobalt from the commandline.

This has been fixed. Metacharacters are now escaped within the command
that launcher Cobalt.

Change-Id: Ic4386b213a0a565f2ba4992517f1b91ff9c83482

Former-commit-id: b354d2d091904c99e48379c1f20beb3fa49cadd2