Let scissor rectangle handle rectangle clipping
When a rectangular viewport is positioned fractionally, Skia
internally mismatches it with the integer-positioned scissor
rectangle, resulting in Skia determining the need for an
explicit rectangle clip in the shader, increasing the set of
shaders we would need to support.

This CL makes Skia not explicitly add a shader stage for
rectangle clipping if it knows that the scissor rectangle will
be enough by itself.

b/27384331

This change is a port of a change originally reviewed at go/lbreview/52282.

Change-Id: I57637dd32073eaeb6339ba27a8decf7dd520b84b

Former-commit-id: 4f2924d315dd6593fae0cbae0b70a1e4ab265828