Fix green right / bottom edge of win32/uwp videos.
This was due to the shader sampling an extra texel along the
borders for bilinear filtering. Using textures of the correct
size will avoid this issue since sampling will be clamped to
the edge. Unfortunately, this means textures will need to be
reallocated when the resolution changes, so there may be a
slight hitch.

Change-Id: I8c5dbf8293772f5c9aa5c5899b9b8eda3edaca64

Former-commit-id: 1e46d75002345a213178feb62077a12826c74ecb