Ensure that pending StorageManager flushes occur on shutdown
Previously, it was possible for pending flushes in the StorageManager to
be lost if the StorageManager destructor was called while an OnChange
timer was active. In this case, the timer would never go off and the
flush would never occur.

This has been fixed by making any pending timers fire as the first step
in the StorageManager destructor, which ensures that the flush occurs.

Tests have been added for this case.

This fixes b/70169619.

Change-Id: I24092fc75229b4852f9e0016d430facd985961d3

Former-commit-id: 7924b079b6fe3c051fc844e5f64276141b0d5572