Always submit new debug console render tree to renderer
There was a bug in OnDebugConsoleRenderTreeProduced() where it would
early out and not submit the render tree to the renderer in the case
where the console had been turned off. However, this resulted in
the console not going away after it was turned off until some other
action caused another render tree to be submitted.

This has been changed so that the render tree is always submitted to the
renderer. The debug console now properly goes away when it is turned
off.

Change-Id: Ia70d2792fac9adba9b1f2d06438158bc920cd496

Former-commit-id: 18bf41d291a37b3ff7b2f8968be2014fffe354a8