Only submit debug console render tree when change occurs
It was possible for multiple render trees to occasionally be produced by
the debug console when it was closed. While this is rare, I've seen it
produce as many as 7 render trees. To prevent the extra ones from
triggering submissions to the renderer, we now check for whether or not
the layer is already empty.

Change-Id: I1526d9d770510661b1be6e19b09f3070e6a26228

Former-commit-id: 0619b08ff2d744ba4ce2797bdc3282b58a1d3f16