Fully initialize storage manager in constructor
Previously, the storage manager would lazily be initialized at some
arbitrary point in the future. This caused issues because the
initialization was not blocking and it meant that cookies could be
retrieved as an upgrade was occurring.

This has been modified so that the initialization fully occurs in the
constructor, guaranteeing that it'll be complete done when any
additional calls are made.

The storage manager tests have temporarily been disabled as a result of
them relying on StorageManager calls being virtual. They'll be rewritten
in a less invasive way.

Change-Id: I5a9cbaf3165c1aeb2d6a1b2d633ca2af1106291d

Former-commit-id: 54bc9aa58a343eb9c523bdb113386c00108f6cfe