StarboardBridge is the owner of all JNI methods.
There are still helper classes with JNI methods, but references to those
helpers are acquired through StarboardBridge in the first place.

While it's still tightly coupled to the CobaltActivity, this sets the
groundwork for changing the life of the StarboardBridge to be associated
with the Application instead of the Activity.

b/69682799

Change-Id: Ie7a1e99c13e56032901bc1d8a682aaad880cd98f

Former-commit-id: d08ae1586166cb39181f9078edda975c999c8a8a