Fix rare FontList crash that Samsung was hitting
There previously was an issue with the ReduceMemory function in
WebModule where it was not clearing out layout boxes. As a result of
this, layout boxes with references to old font lists existed after
the font cache was purged. This ultimately led to a crash from a stale
|character_fallback_typeface_map_| reference in the old font list.

ReduceMemory now has the layout manager purge all layout boxes.
Additionally, the logic in FontCache has been changed so that font lists
with external references are retained during the purge. Instead, their
internals are reset so that they don't hold on to references to fonts.
Lastly, the character fallback typeface maps are no longer purged when
font lists are retained. They simply clear the data within each map
instead, ensuring that they won't go stale on the font lists.

This fixes b/70510314.

Change-Id: Iccf116483bab6c244b222f714c9dd7dc57afa383

Former-commit-id: 774103dc0c3c4d5b0fb673d8aae71d5521e6beb4