Track event's initial render tree rasterization
Previously, the event tracker was unable to track the exact
rasterization time for the first render tree generated by an event. This
was a blind spot in our tracking. The event tracker either had to go
with the time the render tree was created, which didn't include
rasterization, or the start of the animation, which only worked for
events that generate animations. For example, we'd be unable to properly
track the input latency in limited animation mode for tile-to-tile and
shelf-to-shelf events, because they rely upon the animation start time
currently; we're also unable to track the exact time when the black
screen appears while starting a video, because there's no animation tied
to that render tree.

This has been fixed. The WebModule now adds a rasterize callback to
render trees when an event is active. This callback is then used to
record the time that the rasterization occurred, which is provided to
the WebModuleStatTracker, which uses this to record the delay from the
start of the event until the initial render tree is rasterized.

tv_testcase_event_recorder has been updated to record this value and the
DB has been updated to include the new columns.

Change-Id: I34dfa36a28fcd60293e3b4246e3330bfd0ae9818

Former-commit-id: cc538099a6d06d89fb7473a50163c5d4887ab58c