Use parsed switches in CommandLine object for start event.
Don't keep the original char** argv around.

This eliminates the need to maintain and later cleanup a vector of
strings for the virtual command line synthesized from the Android Intent
extras. Instead pass ownership of the strings to the CommandLine, which
itself is passed to the Application to own.

Note: There is still a copy of the strings out of the vector in
CommandLine::InitFromArgv().

Specific changes include:
- Add C++11 move semantics to CommandLine.
- Allow constructing a CommandLine from a vector of strings.
- Pass a CommandLine object to Application::Run() which gets copied or
  moved into the field rather than always creating a new one. Keep
  helper Run() methods that take a char** argv and create the
  CommandLine object.
- Generate a char** argv for the start event from CommandLine just in
  time rather than keeping the original argv. Delete that char** when
  the event is deleted.

Change-Id: Ib44bc6c84520f9689c1ce1c46745adc87b482131

Former-commit-id: 9c5543e76137c3f700377a4b3f468a6bc9dc33b0