Accept Chrome style platform object prototype string conversion
Modify PlatformObjectBindingsTest to accept both the Chrome style and
Firefox style of the resulting string from converting the prototype
object of an interface object.  We accept both of these results because
(e.g.) "String(Object.getPrototypeOf(document.body));" will evaluate to
"[object HTMLBodyElement]" on Chrome 63 and
"[object HTMLBodyElementPrototype]" on Firefox 57.  So, when we use
SpiderMonkey, we handle this in the style of Firefox, and when we use
V8, we handle this in the style of Chrome.  See the bug below for
further details.

b/72125072

Change-Id: I435bd047c227af500a7dfb6869a29f13308a42d9

Former-commit-id: fca9833efbd33605a55de67f54088e97d3bee2e0