Don't re-layout when setting unattached element attributes.
This happens, for example, when the periodic ping timer sets a
"new Image()" object's src attribute to a non-image for ping and
logging purposes.  This used to be causing a re-layout/re-render.
Of course the re-layout was minimal as not much had changed,
however it was still unnecessary, and resulted in a re-rasterization
of the UI.

This was raised as an issue by MStar in this thread:
https://groups.google.com/forum/#!topic/costar/GIke_TnD5e8

Change-Id: I9a53f3731c74e81a231d73384775f030de0bba1e

Former-commit-id: c088804420f49a38157d1e753cdb4f7caa91bc29