Move external Cobalt build from Gradle to CMake
Instead of the brittle hack of depending on the names of the tasks
generated by the Android plugin to determine which ABI to build, now we
just let the plugin invoke CMake and we build the ABI it tells us to.

This allows us to update to Android Studio 3.0.1 since that no longer
generates the per-ABI tasks we relied on, and that also means moving to
Gradle 4.1.

Also update SDK tools and compat libraries to 27.0.2, and GMS to 11.8.0.

b/67501599

Change-Id: I0266340a66585f423ec486ebf8cf9adb1b169d7b

Former-commit-id: f639198fcd3effce6731ff259ec89b79a023fe7c