Disable multiplane image tests (e.g. YUV) for software rasterizer
Previously, we supported rendering of multi plane images like YUV with
the software rasterizer.  However, Skia M61 made some changes that were
incompatible with the shaders we were using to convert YUV images to
RGB.  It was decided that this feature was unlikely to be used going
forward, so it was dropped.

Change-Id: I3b59d4af4a74f1fb0901acf156768facb3d03dd3

Former-commit-id: bb7755b78114d7b4454e4019293d1ce74949b905