Add minor whitespace improvements to interface.cc.template
These are some changes that frequently locally make to the SpiderMonkey
interface.cc.template in order to make it easier to read.  We might as
well check them in.  Essentially, they boil down to treating the jinja
template control lines as more important than the C++ code when it
comes to breaking things up into chunks via newlines.

Change-Id: Iedd15383d4d2e97d95fa4ef4278adb0df856ace5

Former-commit-id: 3c56aba175eeda6619e97ee009fe01ac3667994a