Create starboard launcher for raspi
The old lbshell Raspi launcher has been converted into a starboard
launcher and enabled in the raspi's configuration; this makes it
open-sourceable and usable as a reference by partners.

The logic from run_test.py has also been moved into the raspi launcher.

This CL also fixes ctrl-C on the raspi so that it again works with the
webdriver benchmarks

Change-Id: I7bf8477d6eeac5d395db581aef331040f02cbe19

Former-commit-id: fef4e5fa0b2fddfb774a9d2d710b8ceae56ad2f0