Add workaround for V8 Function name of FunctionTemplate attached to ObjectTemplate quirk
Due to a V8 internals quirks, named constructor attributes MUST come
first, as V8 will use the key that we use to set them on something else
as the "name" property on the function template value, overriding the
originally selected name from |FunctionTemplate::SetClassName|.
Efforts to document/modify this behavior in V8 are underway.

Note that the new context attribute 'all_attributes_v8_order_quirk' is
not immediately used in this CL.  It will be used in the soon to be
checked in larger V8 bindings CL, in which non-static and static
attributes are iterated over together instead of separately.  I felt
this was worth splitting off into a separate CL, so that we don't have
to link the large one to the tracking bug for this quite specific issue.

b/67313140
b/72530786

Change-Id: I6c0b75b9f4a2fe0c1b004fcd8c1dd2850ff33392

Former-commit-id: 0c2c0f6fb011dcbf69ba70e877098933f3b5473f