Remove intptr_t key from |ReferencedObjectMap|
The current API for |ReferencedObjectMap| involves clients calling a
static function to get an intptr_t representation of the |Wrappable*|
they own at the callsite, and then feeding that back to the referenced
object map to add, remove, and use.  I don't think we really get
anything out of this.  I feel that it is both more clear to just pass
in the wrappables, as well as more encapsulated, as the referenced
object map internals can convert it to some other type of key if it
really wants to.

Change-Id: Iebcb5d67db31e0862cc7572a95633075aa239614

Former-commit-id: c1c1cdb85fffb4823b9c0ae997d5bc1a9f9743fa