Add validation of not pseudo class usage in debug and devel
A new define has been added in debug and devel builds,
COBALT_ENABLE_VERSION_COMPATIBILITY_VALIDATIONS. This define enables
all Cobalt version-related validations to run.

The first use case for this is the not pseudo class bug that was
fixed in COBALT 16. When compatibility validations are enabled, the
selector tree checks for any cases of multiple not pseudo classes
existing with the same prefix but different arguments. When this
occurs, it logs a warning, along with the problematic not pseudo
class prefix.

Additionally, a new debug command line switch has been added:
'min_compatibility_version'. When this is provided with an integer, it
is used as the min version to validate against (the pseudo class
validation only occurs when the min version is < 16). When this is not
provided, then validation occurs against all active versions of Cobalt.

When a new violation is encountered, a CVal tracking the total number
of version compatibility violations is incremented and the violation
is logged, with [VERSION COMPATIBILITY VIOLATION] prepended to
facilitate parsing.

This fixes b/72709309 and b/72709444.

Change-Id: Id0456d4f17e13acdfbf50d0ad1aee5ef7da9b3d0

Former-commit-id: d59436ca605b73117e6fb98dac00ac04525bf8d7