Sleep 1ms after logging when running unit tests.
On Android logs are written to a kernel ring buffer, so this avoids
logging so much so fast that the reading process can't process it and
dropping some of the log output.

This replaces the logic that wrote the logs to a local file in addition
to the Android logcat output.

b/38207957

Change-Id: I22ac51345b0a7ed7f475f95f47aaecdca4762a9c

Former-commit-id: 667cd3b18549a3b7fc1d4def694b1cfeb560c4e2