Remove limited_with_jp font package
This package causes bolding anomalies between Japanese glyphs and
Chinese and Korean glyphs, as a result of NotoSansJP-Regular.otf
being synthetically boldable while DroidSansFallback.ttf is not. To
prevent these issues and ensure a more consistent UI, we're removing
this package and the standalone Japanese font. Now, it'll be all or
nothing with having the high quality CJK fonts or the low quality
CJK fonts.

The old limited_with_jp font package now maps to limited.

Change-Id: Idc843c2d756867f09f499970a69d89d76559113f

Former-commit-id: 0d9e763d7797c0876c8c91d9e64a40675d5add84