Improve HTMLMediaElement timeupdate time accuracy
The timeupdate event is required by the spec to fire every 15-250ms.
Our conformance test fails if the interval between timeupdate is more
than 400ms.
The currently timeupdate event is controlled by a timer with an interval
of 250ms but also throttled in the HTMLMediaElement implementation by
checking the absolute interval between two timeupdate events.  Now the
throttling is removed.

b/70309656

Change-Id: I1383ccc5f556ef28aecbb97030558cd0035bfdb1

Former-commit-id: d1ce1330d450aeb51395b5e319ac19e2e2e3200a