Move inject event duration tracking to event dispatch instead
This change moves event duration tracking from injection to dispatch.
This decision was motivated by the fact that injection is not exposed to
JavaScript (it is only exposed to the system, and priviledged
accessors, such as webdriver), however dispatching is.  This makes it
possible for a JavaScript application to traverse itself, and still
cause CVals to be properly updated.

Change-Id: I8252b9fdb4838e79bb6eed696c2e77406bc3564d

Former-commit-id: c61f37513888ed1635c265ea5cc326fdce568788