Modify event duration to include rasterize time
Previously, the total duration of an event did not include the rasterize
time for the first render tree. It instead stopped when the first render
tree was produced. This meant that it didn't match up well to the input
latency, which is what is shown in the dashboards. This has been fixed
so that the rasterize time is included as a duration within the event
and in the event's total time. The webdriver benchmark output will now
better match what appears in the dashboards.

Also, did some stat tracker cleanup so that the non-event totals that
the stat trackers generate are decoupled from the event totals.

Change-Id: I55e75794145660b6c46d7db975dd7b03abdacffe

Former-commit-id: 9d97050a57a9668cc080802f7b6a7d77b898d47c