Pass platform name from gyp runner to gyp
This change has our gyp runner pass platform name to gyp as a gyp
variable, in order for gyp to be able to later expose it to C++ as a
macro.  The initial use case for this is the construction of the V8
startup snapshot cache file name, which must be unique with respect to
binary.

b/67313140
b/72403384

Change-Id: I6e88c9e0b56bd253515344ff55e1979b5a04b9d4

Former-commit-id: 788cff7de1302bf61890ecaf2e0c85748bb7cbcf