Verify the dispatched event in OnStopDispatchEvent
Previously, WebModuleStatTracker was stopping its dispatched event
time on the first OnStopDispatchEvent call that it received. However,
there was no guarantee that this was actually the end of dispatching the
keypress event that started tracking; more likely, it was from another
event triggered by the keypress event.

To guard against this case, the WebModuleStatTracker now stores the
dispatched event that it is tracking, and verifies that the event is the
correct one in OnStopDispatchEvent.

Change-Id: Ia02d2dc5e54c82e668245a59028edcef2e8c1135

Former-commit-id: 89b2fb11a426a9c924ace9ac09a9484822477345