Use www.example.com as a test hostname
Previously, we were using www.yahoo.com as a test hostname.  However,
this was problematic since DNS would return varying number of A records
when queried for this host.  This was confirmed manually on the command
line.

Going forward, we shall use www.example.com instead of www.yahoo.com,
which seems to be more stable in this regard.

Fixes b/73387570

Change-Id: I7a7ce4b5f4037a49e860a2359d8fe1753ab02176

Former-commit-id: 21a569daaf9fcaf9be5da709894aea0635876f90