Collect per-executable static content for deploy targets.
When copying into content from a copy_foo_data.gypi helper, remember
what is copied in the GYP variable |content_deploy_subdirs| so that the
deploy target can collect just the relevant data to package to deploy to
the target device.

copy_web_data.gypi provides a whole action to merge into a target (as
opposed to just a command to merge into an action) so that it is able to
add to |content_deploy_subdirs|.

Use the starboard copy_test_data.gypi that uses the same technique.

In a few cases that had the copy action as part of the test target
itself, separate it out into its own copy target for consistency.
Use a single shared target to copy test data for png_utils_benchmark
and png_utils_test.

b/72971333

Change-Id: I3fef6b8488d96e727571427dfd21504ac2a935f2

Former-commit-id: edce450690d60a07d0280c24ead7c0517d492305