Reorder frame for performance in lib builds.
Landing this change
https://lbshell-internal-review.git.corp.google.com/c/cobalt/+/96210 since
David has left our team.

There was an additional modification to the code that fixes an issue
where the Cobalt UI would not re-render properly when v-sync is turned
off. We now grab our texture ID after we call Submit on the
HardwareRasterizer to ensure that we're getting the texture ID from the
right render target.

Change-Id: Ib9a61d04de5d60438bec4d788f54cf29d104ff2a

Former-commit-id: 7eb4363776dda172804aabb972b65267aa01fed2