Fix crash on exiting when video is being played
PipelineImpl::Suspend() can be called from thread other than the
WebModule thread and it calls PipelineImpl::Stop() internally.  In such
case it is possible that Stop() is called before Start(), which results
in CHECK inside StartTask().

Now loosen the CHECK a bit and check invalid states one by one in
PipelineImpl so we can:
1. Avoid CHECK when Stop() is called before Start().
2. In case if there is another state that causes the CHECK, we can know
   exactly which state it is.

b/73820833

Change-Id: I818938c8bc8e2379621849589260a58471217bc7

Former-commit-id: 70fcff82a258031beee1026d43682647e30628b6