Modify HtmlScriptElement to free string allocations
Previously, HtmlScriptElement never released its Loader after it
completed. This meant that the TextDecoder within the loader always
held on to its string. On Linux and the Raspi, this amounted to
~4.5 MB between the ***REMOVED*** and Player scripts.

Additionally, HtmlScriptElement made its own copy of the strings on
platforms that did not use ref-counted strings and held on to them in
|content_| even after they were executed, which was wasted memory.
On Linux and the Raspi, this amounted to ~2.5 MB between the ***REMOVED*** and
Player scripts.

The logic has been modified so that HtmlScriptElement uses the string
allocation from the TextDecoder, so that it doesn't make a second
unnecessary copy on platforms that don't ref-count strings. It also
immediately releases the memory after it finishes executing the scripts.
The loader itself is now released after it finishes.

In all, this should save ~7MB on platforms that don't use ref-counted
strings (presumably most platforms since its no longer allowed) and
~4.5MB on platforms that do use them.

Change-Id: Ic453db7e46065e2a7b2b8dc1791be81de8753131

Former-commit-id: 66d558029144ba3c02b8ffa1c16278028595aae0