Handle small matrices in |get_scale_factor|
In a previous CL, go/lbreview/96952, a check was added to prevent
floating point inaccuracies for large scaling matrices.  In addition
to checking for closeness of two terms, it should've only returned false
if the terms were large.  The comment seems to imply this, however the
check was missing in the code.

Fixes b/72671168 (again) and fixes b/73798375.

Change-Id: I0509f1cd9886ddfdae4c9262b086d4c0b6f86bc8

Former-commit-id: 865fe95a99aa8aaa6f62dc7b4217b76752e21964