Add mechanism to interrupt synchronous loads during suspend
When a |Suspend| event triggered in Cobalt, the |WebModule|
would cancel all the pending loaders (amongst other things like
releasing graphics resources).

Previously, if a sychronous load was in flight during processing
a script element, any request to suspend would have to wait
on the web module worker thread until the completion of the
synchronous load.  This meant that the suspend operation could
take a very long time to finish.

Now, a mechanism to interrupt the synchronous load has been added,
so that the loader is interrupted, and the rest of the suspend
logic can resume at a much earlier time.

Since the Application lifecycle enforces that the application goes
Started->Pause->Suspend, the Pause event processing also needs to
interrupt any synchronous loads.

Fixes b/74052959

Change-Id: I056eeb434beaa06e8df5712cba0a80e49d563489

Former-commit-id: b21a4a5760d96310cce06ef75d80896710085fa1