Switch CSSDeclaredStyleData to use base::SmallMap
Over 90% of the CSSDeclaredStyleData have a size of 8 or smaller, so a
SmallMap is being used to handle these cases, while map is being used
for the larger cases.

In Linux and the Raspi, where the default bucket size with hash_map is
set to 193, this change saves a large amount of memory (~4MB on Linux
and ~2MB on the Raspi).

Change-Id: I7c9e2fba70092f944a64a092d4b97a77c8613b34

Former-commit-id: 467d9e82e185d556b336ef1839cb5d2d2970614a