Add dynamic load and dispatch layer for ffmpeg.
This adds a layer for dynamically loading and dynamic dispatch of ffmpeg
library calls.

This allows removing of linking against ffmpeg libraries, and the result
version dependency.  The resulting binary should run on systems with
different versions of ffmpeg installed.  If no ffmpeg libraries are
installed, then Cobalt will still run, but audio or video decoding for the
corresponding codecs will fail.

There is still support for using the same code when ffmpeg is linked to the
binary.  This is currently used for the raspberry pi build because our
sysroot is not prepared to allow dynamic loading.

Also add slightly improved error handling in the filter based player worker
and add the qtcreator generator for the raspi builds.

b/74195501

Change-Id: I6196df745fb11b08582ea211d74852e15e26addf

Former-commit-id: 3a603c57f46a0851270cc50427e487ef672f1c4b