Refactor MSE buffer GC based on cached duration
Previously MSE buffer GC is solely determinated by the combined size of
the DecoderBuffers cached without taking the bookkeeping data of the
DecoderBuffers into account.  For videos with many small encoded frames,
it is possible that the bookkeeping data is much larger than the encoded
frames itself, and will cause issues on platforms with limited memory.

b/74520734

Change-Id: I93d5081da583ccf205557af1ede6605287a62435

Former-commit-id: 0ac431ccad3eef25a8843807c13514385bf13049