Fix graphics memory leak with Blitter API.
If a visual element that needed software skia to rasterizer it
had its scale animated at values larger than 1, then the cached
surfaces for that element would not be cleaned up correctly.

This effect apparently did not occur in ***REMOVED***, however it does
occur frequently in the Kids app, which is where MStar noticed the
issue.

This patch was provided by MStar.

b/74735964

Change-Id: Iecb5bcde7fff1d2280fddafafc8fc6b1cfeea6c5

Former-commit-id: 70c6c4070b55f472a19ae9026d1aa8075c45b736