Only install the SDK when we are planning to build.
The configuration class for Android is also initialized
when we run the app launcher, but in that case we are not
intending to actually build anything, so we don't strictly
need to ensure that the SDK is installed.

Only check this when we know we are actually building, e.g.
when we are querying for the build environment variables.

This helps eng prod use the app launcher with Android builds
without setting up the Cobalt source code.

Change-Id: I89974c5cb91328e44a175efa6469d8f2408fa42d

Former-commit-id: 07bdf0fe1a78c351a7aa2183d791f04c4a316420