Remove tracking of XhrResponseData from JS external memory
Tracking of this resulted in a dramatic browse-to-watch regression
across platforms. The reason for this is that it was not previously
being tracked until garbage collection was enabled on it. By adding it
to what was being tracked, but retaining the logic that does not
decrement the total external memory on Mozjs, we ended up
significantly increasing the total number of garbage collections.

Beyond that, tracking XhrResponseData in JS represents a significant
change from previous behavior that could react adversely to platforms
with a GC threshold set to 1MB. As a result of this, I'm removing
tracking of it to mimic previous functionality.

Change-Id: Ib0d01bf293ce0aecc7be1636eb441aec7f9e5bf3

Former-commit-id: def43549419616060884264a3d4a8e5136aac1e3