Revert setting network module thread priority to high.
Setting the priority to high fixed an issue on the PS3 where
it was running out of kernel memory (presumably because sockets
were not being drained quickly enough).  It is reasoned though
that modern operating systems should have more than enough
memory for this, and by making this change we can win back
performance on single core devices.

b/74965639

Change-Id: Idde33036166ae6cb9702d903a2265e9e37220a86

Former-commit-id: 8e7e67b9a21829fde6cac3e863791c8c70a7e028