Add screenshot parameter to Feedback service
This CL adds an optional ArrayBuffer parameter to the screenshot IDL.
This ArrayBuffer can contain a serialized version of an image that is
supported by Android Bitmap (currently PNG, JPEG, WebP, GIF, and BMP).

The image data is the converted into a java Bitmap object, which can
be passed into the feedback service.

This CL also modifies the FeedbackService API to accept a nullable
Bitmap object.

Part of FR: b/66903595

Change-Id: I106ed1f8c43b1c827d6f75dfa8a5513785efdfff

Former-commit-id: 1c8218ca382e433434841f451b3ea67d62a7bf16