Be more robust in choosing to install the Android toolchain.
Re-install the toolchain based on the MD5 sum of sdk_utils.py rather
than a simple timestamp comparison. This allows the tools to be
re-installed with an older configuration in case git is sync'd back to
before a change in how tools are installed.

Prompt the developer running gyp if they are overriding ANDROID_HOME
and/or ANDROID_NDK_HOME since that can break their build if they are in
the habit of ignoring the warning.

Change-Id: I56638a258162d94d0510eea7d5eaa8228e79918a

Former-commit-id: d153888f9a45620914d7290fa17b2523851ca6fd