Add proper tracing support for mutation observer message loop edge case
Motivation:

  (() => {
    var d = document.createElement('div');
    var c = document.createElement('div');
    var obs = new MutationObserver(() => {});
    obs.observe(d, {childList: true});
    d.appendChild(c);
  })();
  // Both JavaScript and Cobalt have dropped all references to the
  // callback.

  // Full GC triggers now.  Note that V8 debug GC is entirely exact.
  window.gc();

  // The message loop task fires, and attempts to dereference the
  // garbage collected callback.

Solution:

Make MutationObserverTaskManager a |Traceable|, as it conceptually owns
|Traceables| (that happen to be JavaScript objects).

Add functionality to let |Traceable|s be roots.  The alternative to this
would have been getting to the MutationObserverTaskManager from window,
but I feel that that is a less elegant solution, as requiring all root
set nodes to be reachable from the global object feels like a bit of a
hack.

b/77760368

Change-Id: Ib651ccff0a4122f00a799c66b499c1422002c132

Former-commit-id: 813c8efc764939b825d54108ab3285d67dec956f