Have skia respect Starboard platforms that insist on GL2.
Skia will now read that the platform has version 2 instead of 3,
if the platform has declared 'gl_type': 'system_gles2' in their
starboard configuration, even if their driver version reports
version 3.

This causes a problem because setting 'gl_type': 'system_gles2'
causes Cobalt to create a GLES 2 graphics context, which it passes
to Skia to use.  But if Skia thinks it is a GLES 3 context (which
it formerly would), it will try to use GLES 3 features which the
driver has disabled for the requested GLES 2 context.

This caused a problem on some Android devices, such as the
Broadcom Elfin.

b/74833235

Change-Id: I8f0d325b07fcb51b32bed0e7a033947fe0fa68b3

Former-commit-id: 3c36c323f180c0512d9af4aa67436e1e385cccdb