Tie wakelock, audio focus, and active MediaSession together.
Introduce the concept of "media focus" that sets a wakelock, requests
audio focus, and activates the MediaSession all in one place.

Previously, when Starboard was suspended we'd only set the MediaSession
inactive, but that left the wakelock and audio focus as they were while
playing. (The wakelock would be lost anyway when the window was
destroyed.)

b/74754046

Change-Id: Ie5ff510cc888d18471d6cae73f6235db580cbc68

Former-commit-id: 21999f0953d5dd16e69f7c8ee51c9db7444569e3