Disable caching of eval being enabled in mozjs
Firefox incorrectly caches whether or not eval is allowed the first time
it is checked. This is a violation of the spec and occasionally breaks
Cobalt's debug console. This code previously existed in mozjs-24, but
didn't make it over with mozjs-45, so I'm just reviving it.

Here's the original CL that @***REMOVED*** added: go/lbreview/53566

Change-Id: I138aa62b8877428c55e7abe0a807136b5bf2ba2e

Former-commit-id: 14fc7935fd347a8452d43dc4ddc382950756d672