Don't release media focus for a transient loss of audiofocus.
Previously we would abandon audiofocus as a side effect of the player
being paused because of the transient loss of audiofocus. That would
cause AudioManager to not call us back when we should regain audiofocus
and resume playback after the transient condition ended.

Removed cases of gaining transient and may-duck audiofocus since we
never request those.

Added a check to pause the player if we fail to gain audiofocus when we
request it. The recommended logic is to not even start the player if we
can't get audiofocus, but HTML5 player is in charge and we just request
audiofocus as a reaction to it starting, so the best we can do is to
pause playback if there's a problem. This should never happen, but we
want to play nice if it does for some reason.

Added a bunch of logging so we can see what's happening with audiofocus
and MediaSession actions controlling the player.

b/73097959

Change-Id: I0c879800e2e9a4794000218de79de3638960865e

Former-commit-id: e75ee89968fe96b0fa4f3aa2e9479a0f4f3ccd26