Use a more flexible server for blackbox tests.
Previously, black box tests used a SimpleHTTPServer to serve assets.
This had several limitations:
1. Launched an external process, which was not easy to control.
2. Hard coded port
  a. Only 1 instance per machine was possible.
  b. May not be available if the port was not cleaned up properly.
3. Current solution was IPv4 only.
4. SimpleHTTPServer is python2 only.

Now, a test can spawn a HTTP server in a thread on demand.  In addition,
the new server will work in IPv6 only environments, and will be easier
to port to python3, when the time comes.  Also, a custom request handler
can be passed, so that tests can finely control the behavior of the
server when a request is received.

Change-Id: Ica4249b6856cde49d14c66afeecdbebf985b7e93

Former-commit-id: 097da7901768370efec94fecdcf68d0ecb549b02