Set the V8 stack size proportional to the web module stack size.
This is following in the lead of MozJS in setting the JavaScript
stack size to 3/4 of the WebModule thread stack size.

In Debug builds, frames tend to be much larger, so it is easier
to run out of stack space.  We already adapt the web module stack
size to be larger in Debug builds, so we should just latch on
to that size to set the V8 JavaScript stack size.

If we don't do this, we end up using the V8 default stack limit
value of 984Kb, defined here:
https://cs.corp.google.com/cobalt/v8/src/globals.h?l=97

Change-Id: Ie751fc24ed4dc5fcaf99909da16f230742cf61f4

Former-commit-id: cfe1abf1e5c8f510c2bd75e7cf967915f75f94d3