Add silence tail to AudioRenderer
AudioRenderer shouldn't rely on the AudioSink to be able to consume very
small amount of audio frames as the AudioSink may have its own warm up
or underflow logic that will start playback only after a certain number
of frames are appended.
EOS is the ideal way to solve this issue but WebAudio logic depends on
this heavily.  So add silence tail to ensure that any amount of audio
frames will be consumed.

b/78791837

Change-Id: Ia03ff9c8c36739b6d7fb7701f23ff4c48ecc7103

Former-commit-id: 5b9ebc174d313841b4deb05dc9d8072363a44e4b