Split SbDecodeTargetPrivate for win32 and uwp.
This splits the SbDecodeTargetPrivate into a base class with separate
specializations for the win32 hardware decoder and XBox One GPU VP9 decoder.

This allows both implementations to be compiled in, and diverge in size and
implementation from another.

This should prevent certain crashes such those in b/78288700, b/78187796
from returning.

Change-Id: I603deda6ede60b7336cb85cf02e96ae6d51e5dc1

Former-commit-id: d9c1cc7bb819811a50f7527ce8fb5e32dc935178